model_training/requirements.txt
transformers
evaluate
torch
scikit-learn
wandb
tqdm
transformers[torch]
torchmetrics
ipykernel
nltk
sentencepiece
sacremoses
python-dotenv